Manually Import Pusher Library
Because libPusher needs WebSocket Library, but WebRTC need this library as well.
Two *.a file will conflict with each other.
The solution is to remove WebSocket from one library. So this library won't be able to build on its own.
The only way to build is to combine two libraries

Only issues is this can only work for ios-device, instead of ios-simulator